+++ lib/generators/helpers.rb
@@ -1,9 +1,9 @@
 # frozen_string_literal: true
 
-require_relative "helpers/migration"
-require_relative "helpers/initializer"
-require_relative "helpers/config"
-require_relative "helpers/values"
-require_relative "helpers/paas"
 require_relative "helpers/alerts"
-require_relative "helpers/outboxfile"
+require_relative "helpers/config"
+require_relative "helpers/initializer"
+require_relative "helpers/items"
+require_relative "helpers/migration"
+require_relative "helpers/paas"
+require_relative "helpers/values"